requirements.txt
# Core dependencies with compatible versions
streamlit>=1.28.0
numpy>=1.21.0
pandas>=1.3.0

# Hugging Face and transformers
transformers>=4.21.0
torch>=1.12.0
tokenizers>=0.13.0

# Sentence transformers (compatible version)
sentence-transformers>=2.2.0

# Vector search
faiss-cpu>=1.7.2

# LangChain (use more recent stable versions)
langchain>=0.0.350
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Additional utilities
scikit-learn>=1.0.0
python-dotenv>=0.19.0
# Streamlit Cloud optimized requirements
streamlit>=1.28.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.12.0 --find-links https://download.pytorch.org/whl/torch_stable.html
transformers>=4.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
python-dotenv>=0.19.0